Fix up the final pieces
